Add files via upload
An online makeup store website project that contains pages for product sections, a contact page, and a page with a form that contains questions. If 80% of them are answered correctly, the person wins a gift.
It also contains a shopping cart that the user can add products to or delete the product. He can also confirm the order and go to the payment page through this cart.
The user will see the current time on the site's home page.